+++ app/actions/_userActions.ts
@@ -61,3 +61,26 @@ export async function finishOnboarding(data: {
     throw new AuthError("Failed to finish onboarding. Please try again later.");
   }
 }
+
+export async function getUserData() {
+  const user = await currentUser();
+
+  if (!user) {
+    throw new AuthError("You are not signed in.");
+  }
+
+  const userData = await prisma.user.findUnique({
+    where: { id: user.id },
+    select: {
+      salt: true,
+      wrappedPrivateKey: true,
+      publicKey: true,
+    },
+  });
+
+  if (!userData) {
+    throw new AuthError("User data not found.");
+  }
+
+  return userData;
+}
